+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Change package type facet to select, and use all filters for search regardless of query presence (UIEH-1527)
 * View/Edit Package Detail Record: Add a Package Display Name field. (UIEH-1501)
 * Add package access filter to package search and provider package search. (UIEH-1526)
+* Create/Edit Custom Package Detail Record: Add a Package display name field. (UIEH-1499)
 
 ## [11.1.1] (https://github.com/folio-org/ui-eholdings/tree/v11.1.1) (2026-05-28)